# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Study Buddy
  version: 1.0.0
  description: Study Budy Api
servers:
  - url: https://localhost:3000
    description: Local development server

  - url: https://studdy-buddy-api-h7kw3.ondigitalocean.app/
    description: Digital Ocean Server
paths:
...
  /friends/requests/{friendId}:
    post:
      summary: Create Friend Request
      operationId: postrequest
      tags:
        - Friends
      security:
        - bearerAuth:
            - read:users
            - public
      parameters:
        - in: path
          name: friendId
          required: true
          schema:
            type: string
          description: The receiver ID
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
  /friends/requests/{requestId}:
    patch:
      summary: Handle Friend Request
      operationId: handlerequest
      tags:
        - Friends
      security:
        - bearerAuth:
            - read:users
            - public
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
          description: The request ID

      requestBody:
        content:
          application/json:
            schema:
              properties:
                isAccepted:
                  type: boolean

              required:
                - isAccepted
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"

    delete:
      summary: Delete Friend Request
      operationId: deleterequest
      tags:
        - Friends
      security:
        - bearerAuth:
            - read:users
            - public
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
          description: The request ID
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
  /friends/requests:
    get:
      summary: Get Friend Requests
      operationId: getrequests
      tags:
        - Friends
      security:
        - bearerAuth:
            - read:users
            - public

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"

        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
  /friends:
    get:
      summary: Get all Friends of User
      operationId: getAllUserFriends
      tags:
        - Friends

      security:
        - bearerAuth:
            - read:users
            - public

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friend"

        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
  /friends/{friendId}:
    delete:
      summary: Remove Friend
      operationId: removeFriend
      tags:
        - Friends

      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: friendId
          required: true
          schema:
            type: string
          description: The ID of friend being removed

      responses:
        "200":
          description: Success
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
...
  /friend/{friendId}/message:
    post:
      summary: Send a message to a friend
      operationId: sendMessageToFriend
      tags:
        - Messages

      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: friendId
          required: true
          description: ID of the friend to send a message to
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: "hello?"
              required:
                - content
      responses:
        "200":
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    example: "hello"
                  sender:
                    type: string
                    description: The ID of the user sending the message
                  createdAt:
                    type: string
                    format: date-time
        "400":
          description: Invalid friend ID or bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /friend/{friendId}/messages:
    get:
      summary: Get messages exchanged with a friend
      operationId: getMessagesFromFriend
      tags:
        - Messages
      security:
        - bearerAuth:
            - read:users
            - public

      parameters:
        - in: path
          name: friendId
          required: true
          description: ID of the friend whose messages are being retrieved
          schema:
            type: string

      responses:
        "200":
          description: Successfully retrieved messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    sender:
                      type: string
                    content:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
        "400":
          description: Invalid friend ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized
        "500":
          description: Server error
...
components:
  schemas:
    User:
      type: object
      required:
        - firstName
        - lastName
        - username
        - email
        - password
      properties:
        firstName:
          type: string

        lastName:
          type: string

        username:
          type: string
          minLength: 4

        email:
          type: string

        password:
          type: string
          minLength: 8

        friends:
          type: array
          items:
            $ref: "#/components/schemas/Friend"

        incomingRequests:
          type: array
          items:
            $ref: "#/components/schemas/Request"

        outgoingRequests:
          type: array
          items:
            $ref: "#/components/schemas/Request"

        chatSessions:
          type: array
          items:
            $ref: "#/components/schemas/Chat"

      example:
        firstName: "Cornelius"
        lastName: "Bartels"
        email: "cjkb@gmail.com"
        password: "test121111"
    BadRequestError:
      type: object
      properties:
        name:
          type: string
        errors:
          type: array
          items:
            type: object
...
    Request:
      type: object
      required:
        - sender
        - receiver
      properties:
        sender:
          $ref: "#/components/schemas/User"
        receiver:
          $ref: "#/components/schemas/User"
        isAccepted:
          type: boolean
    Friend:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    Chat:
      type: object
      properties:
        chatType:
          type: string
          enum:
            - direct
            - group
        messageBuckets:
          type: array
          items:
            $ref: "#/components/schemas/User"
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        groupName:
          type: string
        owner:
          $ref: "#/components/schemas/User"

      required:
        - chatType
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt